+++ packages/terra-theme-provider/src/ThemeProvider.jsx
@@ -1,5 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import className from 'classnames';
+
+const cx = className;
 
 const ThemeProviderThemes = {
   CONSUMER: 'cerner-consumer-theme',
@@ -53,7 +56,7 @@ class ThemeProvider extends React.Component {
 
   render() {
     const { themeName, isGlobalTheme, children, ...customProps } = this.props;
-    return (<div {...customProps} className={themeName}>{children}</div>);
+    return (<div {...customProps} className={cx(themeName, customProps.className)}>{children}</div>);
   }
 }